requirements.txt
openai
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
streamlit
tiktoken
python-dotenv
numpy
